requirements.txt
gooddata_sdk==1.36.0
openai==1.82.1
pdf2image==1.17.0
python-dotenv==1.1.0
